Fixed bugs, moved to Emacs, connected clangd and clang-format, updated CMakeLists, added rudimentary CMakePresets

+++ src/lab06/main.cxx
@@ -1,6 +1,6 @@
+#include <filesystem>
 #include <fstream>
 #include <ostream>
-#include <filesystem>
 
 #include <fmt/color.h>
 #include <fmt/ostream.h>
@@ -10,23 +10,21 @@
 #include "lab06.h"
 
 
-template <std::floating_point DType>
+template<std::floating_point DType>
 [[nodiscard]]
-auto read_input_file(const std::string& filename) -> Outlab6<DType> {
-    const auto input_filepath = std::filesystem::path{ filename };
+auto read_input_file(const std::string &filename) -> Outlab6<DType> {
+    const auto input_filepath = std::filesystem::path{filename};
 
     if (input_filepath.empty()) {
         throw std::runtime_error(
-            fmt::format("Input file does not exist: {}", input_filepath.string())
-        ); // Indicate an error occurred
+                fmt::format("Input file does not exist: {}", input_filepath.string())); // Indicate an error occurred
     }
 
-    std::ifstream input{ input_filepath };
+    std::ifstream input{input_filepath};
 
     if (!input.is_open()) {
         throw std::runtime_error(
-            fmt::format("Could not open input file: {}", input_filepath.string())
-        ); // Indicate an error occurred
+                fmt::format("Could not open input file: {}", input_filepath.string())); // Indicate an error occurred
     }
 
     try {
@@ -34,29 +32,27 @@ auto read_input_file(const std::string& filename) -> Outlab6<DType> {
         input.close();
         return lab;
     }
-    catch (const std::exception&) {
+    catch (const std::exception &) {
         input.close();
         throw;
     }
 }
 
 
-using real = long double;
+using real = double;
 
-int main(int argc, char** argv) {
+int main(int argc, char **argv) {
     const Header header{};
 
     argparse::ArgumentParser program{
-        "shumilov_outlab6",
-        "1.0",
-        argparse::default_arguments::help,
+            "shumilov_outlab6",
+            "1.0",
+            argparse::default_arguments::help,
     };
 
-    program.add_argument("input")
-           .help("Path to input file");
+    program.add_argument("input").help("Path to input file");
 
-    program.add_argument("-o", "--output")
-           .help("Path to output file");
+    program.add_argument("-o", "--output").help("Path to output file");
 
     try {
         program.parse_args(argc, argv);
@@ -64,14 +60,11 @@ int main(int argc, char** argv) {
 
         const auto result = lab.run();
 
-        if (const auto output_filename = program.present<std::string>("--output");
-            output_filename.has_value()
-        ) {
-            std::ofstream output{ output_filename.value() };
+        if (const auto output_filename = program.present<std::string>("--output"); output_filename.has_value()) {
+            std::ofstream output{output_filename.value()};
             if (!output.is_open()) {
                 throw std::runtime_error(
-                    fmt::format("Could not open: '{}'", output_filename.value())
-                ); // Indicate an error occurred
+                        fmt::format("Could not open: '{}'", output_filename.value())); // Indicate an error occurred
             }
             header.echo(output);
             result.echo(output);
@@ -81,13 +74,10 @@ int main(int argc, char** argv) {
             result.echo(std::cout);
         }
     }
-    catch (const std::exception& err) {
-        std::cerr << "\n"
-          << format(fmt::emphasis::bold | fg(fmt::color::red), "Error: ")
-          << err.what() << "\n\n";
+    catch (const std::exception &err) {
+        std::cerr << "\n" << format(fmt::emphasis::bold | fg(fmt::color::red), "Error: ") << err.what() << "\n\n";
         std::exit(EXIT_FAILURE);
     }
 
     return EXIT_SUCCESS;
-
 }
